Make simple matcher into struct
... and move non methods to util file
